+++ src/FEValues/CellValues.jl
@@ -123,7 +123,7 @@ function reinit!(cv::AbstractCellValues, cell::Union{AbstractCell, Nothing}, x::
 
     check_reinit_sdim_consistency(cv, x)
     if cell === nothing && reinit_needs_cell(cv)
-        throw(ArgumentError("The cell::AbstractCell input is required to reinit! non-identity function mappings"))
+        throw(ArgumentError("The cell::AbstractCell input is required to reinit! this function mapping"))
     end
     if !checkbounds(Bool, x, 1:n_geom_basefuncs) || length(x) != n_geom_basefuncs
         throw_incompatible_coord_length(length(x), n_geom_basefuncs)
@@ -132,6 +132,7 @@ function reinit!(cv::AbstractCellValues, cell::Union{AbstractCell, Nothing}, x::
         mapping = calculate_mapping(geo_mapping, q_point, x)
         _update_detJdV!(getdetJdVs(cv), q_point, w, mapping)
         apply_mapping!(fun_values, q_point, mapping, cell)
+        apply_dof_transformation!(fun_values, q_point, mapping, cell)
     end
     return nothing
 end
@@ -274,7 +275,12 @@ get_quadrature_rule(cv::MultiFieldCellValues) = getfield(cv, :qr)
 end
 
 @inline function reinit_needs_cell(cv::MultiFieldCellValues)
-    return any(map(fv -> !isa(mapping_type(fv), IdentityMapping), get_fun_values(cv)))
+    return any(
+        map(
+            fv -> mapping_needs_cell(mapping_type(fv)) || dof_transformation_needs_cell(dof_transformation(fv)),
+            get_fun_values(cv)
+        )
+    )
 end
 
 function check_reinit_sdim_consistency(cmv::MultiFieldCellValues, ::AbstractVector{VT}) where {VT}
@@ -287,6 +293,17 @@ end
     expr = Expr(:block)
     for i in 1:N
         push!(expr.args, :(apply_mapping!(fun_values[$i], q_point, mapping, cell)))
+    end
+    return quote
+        $(Expr(:meta, :inline))
+        @inbounds return $expr
+    end
+end
+
+@generated function apply_dof_transformation!(fun_values::Tuple{Vararg{FunctionValues, N}}, q_point, mapping, cell) where {N}
+    expr = Expr(:block)
+    for i in 1:N
+        push!(expr.args, :(apply_dof_transformation!(fun_values[$i], q_point, mapping, cell)))
     end
     return quote
         $(Expr(:meta, :inline))